fix filter reservations cost and integer

+++ database/migrations/2017_09_17_135346_create_reservations_table.php
@@ -15,7 +15,7 @@ class CreateReservationsTable extends Migration
     {
         Schema::create('reservations', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('cost');
+            $table->float('cost');
             $table->boolean('status')->default(false);
             $table->boolean('payment_status')->default(false);
